Add ga event for halfway reads
Firing an event off when the visitor scrolls halfway through the article
content will tell GA to not count the visit as a bounce if that visitor
ends up not visiting any another links.

Recommended suggestions from GA on setting a timeout were pretty shitty
compared to going this route, inspired from a blog post on LinkedIn.